vpa to uploader

+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/archiving/adapters/vpa/ArchiveSourceAdapterFolder.java
@@ -3,16 +3,19 @@ package de.mephisto.vpin.server.archiving.adapters.vpa;
 import de.mephisto.vpin.restclient.archiving.ArchivePackageInfo;
 import de.mephisto.vpin.restclient.archiving.ArchiveType;
 import de.mephisto.vpin.restclient.frontend.TableDetails;
+import de.mephisto.vpin.restclient.util.FileUtils;
 import de.mephisto.vpin.server.archiving.ArchiveDescriptor;
 import de.mephisto.vpin.server.archiving.ArchiveSource;
 import de.mephisto.vpin.server.archiving.ArchiveSourceAdapter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.awt.*;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.util.*;
+import java.util.List;
 
 public class ArchiveSourceAdapterFolder implements ArchiveSourceAdapter {
   private final static Logger LOG = LoggerFactory.getLogger(ArchiveSourceAdapterFolder.class);
@@ -66,9 +69,13 @@ public class ArchiveSourceAdapterFolder implements ArchiveSourceAdapter {
   @Override
   public boolean delete(ArchiveDescriptor descriptor) {
     File file = new File(archiveFolder, descriptor.getFilename());
-    if (!file.delete()) {
-      LOG.error("Failed to delete " + file.getAbsolutePath());
-      return false;
+    LOG.info("Deleting {}", file.getAbsolutePath());
+    if (!Desktop.getDesktop().moveToTrash(file)) {
+      LOG.error("Failed moving file to trash: " + file.getAbsolutePath());
+      if (!file.delete()) {
+        LOG.error("Failed to delete " + file.getAbsolutePath());
+        return false;
+      }
     }
     this.invalidate();
     return true;
